requirements.txt
streamlit
pypdf
Pillow
reportlab
python-docx